backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-environ==0.11.2
psycopg[binary]==3.2.3
celery[redis]==5.4.0
redis==5.2.0
django-celery-beat==2.7.0
gunicorn==23.0.0
